requirements.txt
flask
flask-sqlalchemy
flask-cors
pyjwt
werkzeug
razorpay
gunicorn
python-dotenv
